more compression, auto added refresh
Took out alot of stuff and then put it back, some things may have stayed the same, or not. so far not sure if the auto refresh is working, i have it set to refresh every 5 seconds so now just have to wait and see when someone comes online with out me refresing the page manualy.